연산 로직 ArithmeticCalculator.java로 이동, 제네릭 사용해 double도 수용 가능

+++ src/Lv3/OperatorType.java
@@ -23,7 +23,7 @@ public enum OperatorType {
         return null;
     }
 
-    public long calculate(long x, long y){
+    public double calculate(Double x, Double y){
         if(PLUS.getOperator() == this.getOperator()){
             return x + y;
         }
